ExecutionDisplay refactoring
- Extract logger class
- Extract SummaryDisplay class
- Extract ExecutedSpecs
- Move responsabilities to SpecReporter
- Reorganize sources
