+++ examples/typescript/README.md
@@ -4,8 +4,8 @@ Use jasmine-spec-reporter with TypeScript
 ## Configuration
 
 ```typescript
-import { SpecReporter } from "jasmine-spec-reporter";
-import { DisplayProcessor } from "jasmine-spec-reporter";
+import {SpecReporter} from "jasmine-spec-reporter";
+import {DisplayProcessor} from "jasmine-spec-reporter";
 const Jasmine = require("jasmine");
 import SuiteInfo = jasmine.SuiteInfo;
