--- src/custom-reporter-result.ts
@@ -1,3 +1,0 @@
-export interface CustomReporterResult extends jasmine.CustomReporterResult {
-    duration?: string;
-}
