+++ src/main.ts
@@ -1,2 +1,2 @@
-export { SpecReporter } from "./spec-reporter";
-export { DisplayProcessor } from "./display/display-processor";
+export {SpecReporter} from "./spec-reporter";
+export {DisplayProcessor} from "./display-processor";
